feat: add dynamic routing
