+++ package.json
@@ -15,7 +15,8 @@
     "date-fns": "^4.1.0",
     "lucide-react": "^0.488.0",
     "react": "^19.0.0",
-    "react-dom": "^19.0.0"
+    "react-dom": "^19.0.0",
+    "react-router": "^7.5.0"
   },
   "devDependencies": {
     "@eslint/js": "^9.24.0",
