+++ src/App.jsx
@@ -1,12 +1,18 @@
 import React from 'react'
-import Conversation from './components/conversation/conversation'
-import Sidebar from './components/sidebar/sidebar'
+import { createBrowserRouter, RouterProvider } from 'react-router'
+import DefaultPage from './components/conversation/components/default-page'
+import RootLayout from './pages/Root'
+
+const router = createBrowserRouter([
+  {
+    path: '/',
+    element: <RootLayout />,
+    children: [
+      { path: "/", element: <DefaultPage /> },
+    ]
+  }
+])
 
 export default function App() {
-  return (
-    <div className='flex'>
-      <Sidebar />
-      <Conversation />
-    </div>
-  )
+  return <RouterProvider router={router} />
 }
