+++ src/pages/Root.jsx
@@ -1,0 +1,12 @@
+import React from 'react'
+import Conversation from '../components/conversation/conversation'
+import Sidebar from '../components/sidebar/sidebar'
+
+export default function RootLayout() {
+    return (
+        <div className='flex'>
+            <Sidebar />
+            <Conversation />
+        </div>
+    )
+}
